Add build and config files for PGO feature testing
Profile-guided optimization (PGO) is a technique to improve the runtime
performance of a program by relying on previously collected information.
It runs in two phases:
1. the generation phase: the compiler and linker instrument the code to
collect profiling information at runtime
2. the optimization phase: the compiler uses the information collected
in the generation phase to optimize the program

This Unikraft application is simply printing "Hello, World!" to standard
output.  It's not meant to make use of profile-guided optimization
(PGO), but merely to showcase its configuration and use in Unikraft.

It uses both kraft and the manual approach (Makefile-based) to build and
run the Unikraft application.

Signed-off-by: Razvan Deaconescu <razvan.deaconescu@cs.pub.ro>
Reviewed-by: Cezar Craciunoiu <cezar.craciunoiu@gmail.com>
Reviewed-by: Sergiu Moga <sergiu.moga@protonmail.com>
Reviewed-by: Alice Suiu <alicesuiu17@gmail.com>