Add build and config files for PGO feature testing
It uses the old-style Makefile-based (i.e. non-kraft) way of building
the application.

A config.sample file is provided to validate the configuration done via
`make menuconfig`.

The application is simply printing "Hello, World!" to standard output.
It's not meant to make use of profile-guided optimization (PGO), but
merely to showcase its configuration and use in Unikraft.

Signed-off-by: Razvan Deaconescu <razvan.deaconescu@cs.pub.ro>